+++ app/lib/widgets/main/navigation_bar.dart
